etterlatte-common Flyttet litt data classes til common
