+++ apps/etterlatte-grunnlag/src/main/resources/db/migration/V4__legg_til_fnr_grunnlagshendelse.sql
@@ -1,2 +1,2 @@
 ALTER TABLE grunnlagshendelse
-ADD fnr text
+ADD if not exists fnr text
