+++ apps/etterlatte-opplysninger-fra-pdl/src/main/kotlin/opplysninger/kilde/pdl/BesvarOpplysningsbehov.kt
@@ -3,6 +3,7 @@ package no.nav.etterlatte.opplysninger.kilde.pdl
 import no.nav.etterlatte.common.objectMapper
 import no.nav.etterlatte.libs.common.grunnlag.opplysningstyper.Opplysningstyper
 import no.nav.etterlatte.libs.common.logging.withLogContext
+import no.nav.etterlatte.libs.common.pdl.PersonDTO
 import no.nav.etterlatte.libs.common.person.Foedselsnummer
 import no.nav.etterlatte.libs.common.person.Person
 import no.nav.etterlatte.libs.common.person.PersonRolle
