+++ libs/common/src/main/kotlin/grunnlag/opplysningstyper/Navn.kt
@@ -1,0 +1,6 @@
+package no.nav.etterlatte.libs.common.grunnlag.opplysningstyper
+
+data class Navn(
+    val fornavn: String,
+    val etternavn: String
+)
